feat: Refactor GUI components and rendering modules
