+++ src/main/java/io/github/waqfs/gui/hud/notification/NotificationDisplay.java
@@ -7,6 +7,8 @@ import io.github.waqfs.Cigarette;
 import io.github.waqfs.gui.CigaretteScreen;
 import io.github.waqfs.gui.hud.notification.internal.NotificationWithEasingProgress;
 import io.github.waqfs.gui.widget.DraggableWidget;
+import io.github.waqfs.lib.Color;
+import io.github.waqfs.lib.Shape;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.font.TextRenderer;
 
@@ -62,7 +64,7 @@ public class NotificationDisplay extends ClickableWidget {
 
         List<Integer> colors = new ArrayList<>();
         for (int i = 0; i < "cigarette".length(); i++) {
-            colors.add(DraggableWidget.color(x + i * 10, y + 40));
+            colors.add(Color.color(x + i * 10, y + 40));
         }
         drawTextGradient(context, "cigarette", x, y + 40, colors.stream().mapToInt(i -> i).toArray());
     }
@@ -73,7 +75,7 @@ public class NotificationDisplay extends ClickableWidget {
                 Cigarette.LOGO_IDENTIFIER, x + 7, y, 0f, 0f, 30, 30, 30, 30);
         List<Integer> colors = new ArrayList<>();
         for (int i = 0; i < "cigarette".length(); i++) {
-            colors.add(DraggableWidget.color((int) (x + i * 10 * gradientStaggerModifier),
+            colors.add(Color.color((int) (x + i * 10 * gradientStaggerModifier),
                     (int) (y + 40 * gradientStaggerModifier)));
         }
         drawTextGradient(context, "cigarette", x, y + 40, colors.stream().mapToInt(i -> i).toArray());
@@ -190,7 +192,7 @@ public class NotificationDisplay extends ClickableWidget {
                                     : notificationType.equals("error") ? 0xFFFF5C5C : 0xFF999999);
 
             int bg = CigaretteScreen.BACKGROUND_COLOR;
-            DraggableWidget.roundedRect(context, clampedLeft, clampedTop, clampedRight, clampedBottom, bg,
+            Shape.roundedRect(context, clampedLeft, clampedTop, clampedRight, clampedBottom, bg,
                     cornerRadius);
 
             float visibleStart = NotificationWithEasingProgress.APPEAR_TICKS;
@@ -220,7 +222,7 @@ public class NotificationDisplay extends ClickableWidget {
                         cornerRadius,
                         Math.min((barRight - barLeft) / 2, barHeight / 2)));
 
-                DraggableWidget.roundedRect(context,
+                Shape.roundedRect(context,
                         barLeft + 1, barTop, barRight, barBottom,
                         0xFF00FF00,
                         0,
