+++ src/main/java/io/github/waqfs/gui/widget/ScrollableWidget.java
@@ -2,6 +2,7 @@ package io.github.waqfs.gui.widget;
 
 import io.github.waqfs.gui.CigaretteScreen;
 import io.github.waqfs.gui.Scissor;
+import io.github.waqfs.lib.Shape;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.widget.ClickableWidget;
 import net.minecraft.text.Text;
@@ -273,7 +274,7 @@ public class ScrollableWidget<Widgets extends BaseWidget<?>>
                 }
                 int bottomRectTop = realBottomInt;
                 if (this.getEasedProgress() > 0.0 && showBottomRoundedRect) {
-                    DraggableWidget.roundedRect(context, left, bottomRectTop, right,
+                    Shape.roundedRect(context, left, bottomRectTop, right,
                             bottomRectTop + BOTTOM_ROUNDED_RECT_HEIGHT,
                             CigaretteScreen.BACKGROUND_COLOR, 5, false, true);
                 }
