+++ src/main/java/io/github/waqfs/module/ui/Notifications.java
@@ -1,9 +1,7 @@
-package io.github.waqfs.module.render;
-
-import org.jetbrains.annotations.NotNull;
+package io.github.waqfs.module.ui;
 
 import io.github.waqfs.Cigarette;
-import io.github.waqfs.gui.hud.modules.ModuleListDisplay;
+import io.github.waqfs.gui.hud.notification.NotificationDisplay;
 import io.github.waqfs.gui.widget.ToggleWidget;
 import io.github.waqfs.module.RenderModule;
 import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderContext;
@@ -11,26 +9,26 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.client.world.ClientWorld;
+import org.jetbrains.annotations.NotNull;
 
-public class ModuleList extends RenderModule<ToggleWidget, Boolean> {
-    protected static final String MODULE_NAME = "Module List";
-    protected static final String MODULE_TOOLTIP = "Displays a list of modules.";
-    protected static final String MODULE_ID = "render.module_list";
+public class Notifications extends RenderModule<ToggleWidget, Boolean> {
+    protected static final String MODULE_NAME = "Notifications";
+    protected static final String MODULE_TOOLTIP = "Displays notifications.";
+    protected static final String MODULE_ID = "render.notifications";
 
-    private ModuleListDisplay display;
+    private NotificationDisplay display;
 
-    public ModuleList() {
+    public Notifications() {
         super(ToggleWidget::module, MODULE_ID, MODULE_NAME, MODULE_TOOLTIP);
-        this.setChildren();
         this.widget.withDefaultState(true);
     }
 
     @Override
     protected void onWorldRender(WorldRenderContext ctx, @NotNull MatrixStack matrixStack) {
         if (display == null) {
-            display = new ModuleListDisplay();
-            Cigarette.registerHudElement(display);
+            display = new NotificationDisplay();
         }
+        Cigarette.NOTIFICATION_DISPLAY = display;
     }
 
     @Override
@@ -40,10 +38,9 @@ public class ModuleList extends RenderModule<ToggleWidget, Boolean> {
 
     @Override
     protected void onDisabledTick(MinecraftClient client) {
-
-        if (display != null) {
-            Cigarette.unregisterHudElement(display);
-            display = null;
+        if (Cigarette.NOTIFICATION_DISPLAY == display) {
+            Cigarette.NOTIFICATION_DISPLAY = null;
         }
+        display = null;
     }
 }
